refactor(workflows): own session replay reducer
Generated-By: pi 0.84.3 (https://pi.dev)
